refactor: extract duplicated types and helpers to shared modules
- ColumnAlign type + makeAlignClasses() → shared/table.ts
- formatFL() → utils/formatting.ts
- makeLens() test factory → test/factories.ts

Each duplicated item now exists in exactly one place.

Closes #290
